[HUDI-2941] Show _hoodie_operation in spark sql results
